worked on checkoutcart

+++ Webshoppen2/Webshoppen2/Models/CheckoutCart.cs
@@ -9,7 +9,7 @@ namespace Webshoppen2.Models
     internal class CheckoutCart
     {
         public int Id { get; set; }
-        public int OrderId { get; set; }
+        public float OrderId { get; set; }
         public int? ProductId { get; set; }
         public int? AmountofUnits { get; set; }
         public int? CustomerId { get; set; }
